% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_download_file}
\alias{czso_download_file}
\title{Download file from blah}
\usage{
czso_download_file(url, dest_dir = NULL)
}
\arguments{
\item{url}{URL to download from. Must be character of length one. A result of \code{czso_get_dataset_url()}.}

\item{dest_dir}{character. Directory in which downloaded files will be stored.
If left unset, will use the \code{czso.dest_dir} option if the option is set, and \code{tempdir()} otherwise. Will be created if it does not exist.}
}
\value{
Path to downloaded file (character of length one)
}
\description{
Download file from blah
}
\examples{
print("blah")
}
\seealso{
Other Fine-grained control workflow: 
\code{\link{czso_download_if_needed}()},
\code{\link{czso_read_csv}()}
}
\concept{Fine-grained control workflow}
