% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_download_if_needed}
\alias{czso_download_if_needed}
\title{Title}
\usage{
czso_download_if_needed(url, dest_dir = NULL, force_redownload)
}
\arguments{
\item{url}{Blah}

\item{dest_dir}{Blah}

\item{force_redownload}{Blah}
}
\value{
Blah
}
\description{
Title
}
\examples{
"blah"
}
\seealso{
Other Fine-grained control workflow: 
\code{\link{czso_download_file}()},
\code{\link{czso_read_csv}()}
}
\concept{Fine-grained control workflow}
