% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_get_dl_path}
\alias{czso_get_dl_path}
\title{Title}
\usage{
czso_get_dl_path(url, dest_dir = NULL)
}
\arguments{
\item{url}{Blah}

\item{dest_dir}{Blah}
}
\value{
Thing
}
\description{
Title
}
\examples{
print("blah")
}
