% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_get_url}
\alias{czso_get_url}
\title{Title}
\usage{
czso_get_url(dataset_id = NULL, resource_num = 1)
}
\arguments{
\item{dataset_id}{a character. Found in the czso_id column of data frame returned by \code{get_catalogue()}.}

\item{resource_num}{integer. Order of resource in resource list for the given dataset. Defaults to 1, the normal value for CZSO datasets.}
}
\value{
Thing
}
\description{
Title
}
\examples{
print("Blah")
}
