% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_read_csv}
\alias{czso_read_csv}
\title{Title}
\usage{
czso_read_csv(path)
}
\arguments{
\item{path}{}
}
\value{
Blah
}
\description{
Title
}
\examples{
print("blah")
}
\seealso{
Other Fine-grained control workflow: 
\code{\link{czso_download_file}()},
\code{\link{czso_download_if_needed}()}
}
\concept{Fine-grained control workflow}
