% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{czso_target_table}
\alias{czso_target_table}
\title{A target factory for using CZSO data in {targets} pipelines}
\usage{
czso_target_table(name, dataset_id, dest_dir, resource_num = 1, freeze = FALSE)
}
\arguments{
\item{name}{DESCRIPTION.}

\item{dataset_id}{DESCRIPTION.}

\item{dest_dir}{DESCRIPTION.}

\item{resource_num}{DESCRIPTION.}

\item{freeze}{Whether to track the}
}
\value{
a list of three targets, named \verb{\{name\}_url}, \verb{\{name\}_file} and \verb{\{name\}_data}.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
