% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_catalogue}
\alias{get_catalogue}
\title{Get catalogue of open CZSO datasets}
\usage{
get_catalogue(
  provider = "<U+010C>esk<U+00FD> statistick<U+00FD> <U+00FA><U+0159>ad",
  title_regex = NULL,
  description_regex = NULL
)
}
\arguments{
\item{provider}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
