% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_czso_dataset_doc}
\alias{get_czso_dataset_doc}
\title{Deprecated: use \code{czso_get_dataset_doc()} instead}
\usage{
get_czso_dataset_doc(
  dataset_id,
  action = c("return", "open", "download"),
  destfile = NULL,
  format = c("html", "pdf", "word")
)
}
\arguments{
\item{dataset_id}{Dataset ID}

\item{action}{Whether to \code{return} URL (the default), \code{download} the file, or \code{open} the URL in the default web browser.}

\item{destfile}{Where to save the file. Only used if if \code{action = download}.}

\item{format}{What file format to access: \code{html} (the default), \code{pdf}, or \code{word}.}
}
\value{
a list
}
\description{
\lifecycle{deprecated}
}
\seealso{
Other Additional tools: 
\code{\link{czso_get_dataset_doc}()},
\code{\link{czso_get_dataset_metadata}()},
\code{\link{czso_get_table_schema}()},
\code{\link{get_czso_dataset_metadata}()},
\code{\link{get_czso_table_schema}()}
}
\concept{Additional tools}
