% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_czso_dataset_metadata}
\alias{get_czso_dataset_metadata}
\title{Get dataset metadata}
\usage{
get_czso_dataset_metadata(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset ID}
}
\value{
a list
}
\description{
Get metadata from CZSO API, which can be somewhat more detailed/readable than
what is provided in the dataset's entry in the output of \code{get_czso_dataset()}.
}
\examples{
get_czso_dataset_metadata("110080")
}
\seealso{
Other Additional tools: 
\code{\link{get_czso_dataset_doc}()},
\code{\link{get_czso_table_schema}()}
}
\concept{Additional tools}
