% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_czso_table_schema}
\alias{get_czso_table_schema}
\title{Deprecated: use \code{czso_get_table_schema()} instead}
\usage{
get_czso_table_schema(dataset_id, resource_num)
}
\arguments{
\item{dataset_id}{Dataset ID}

\item{resource_num}{Resource number, typically 1 in CZSO (the default)}
}
\value{
a list
}
\description{
\lifecycle{deprecated}
}
\seealso{
Other Additional tools: 
\code{\link{czso_get_dataset_doc}()},
\code{\link{czso_get_dataset_metadata}()},
\code{\link{czso_get_table_schema}()},
\code{\link{get_czso_dataset_doc}()},
\code{\link{get_czso_dataset_metadata}()}
}
\concept{Additional tools}
