% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_table}
\alias{get_table}
\title{FUNCTION_TITLE}
\usage{
get_table(dataset_id, resource_num = 1)
}
\arguments{
\item{dataset_id}{DESCRIPTION.}

\item{resource_num}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
