% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czso-package.R
\docType{package}
\name{czso-package}
\alias{czso}
\alias{czso-package}
\title{czso: Use Open Data from the Czech Statistical Office in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Get programmatic access to the open data provided by the
    Czech Statistical Office (CZSO).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petrbouchal/czso}
  \item Report bugs at \url{https://github.com/petrbouchal/czso/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

}
\keyword{internal}
