% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_get_catalogue}
\alias{czso_get_catalogue}
\title{Get catalogue of open CZSO datasets}
\usage{
czso_get_catalogue()
}
\value{
a data frame with details on all CZSO datasets available in the Czech National Open Data Catalogue.
}
\description{
Retrieves a list of all CZSO's open datasets available from the Czech Open data catalogue.
}
\details{
Use the dataset_id column as an argument to \code{get_czso_table()}.
}
\examples{
\donttest{
czso_get_catalogue()
}
}
\seealso{
Other Core workflow: 
\code{\link{czso_get_table}()},
\code{\link{get_table}()}
}
\concept{Core workflow}
