% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_get_table}
\alias{czso_get_table}
\title{Retrieve and read dataset from CZSO}
\usage{
czso_get_table(dataset_id, force_redownload = FALSE, resource_num = 1)
}
\arguments{
\item{dataset_id}{a character. Found in the czso_id column of data frame returned by \code{get_catalogue()}.}

\item{force_redownload}{integer. Whether to redownload data source file even if already cached. Defaults to FALSE.}

\item{resource_num}{integer. Order of resource in resource list for the given dataset. Defaults to 1, the normal value for CZSO datasets.}
}
\value{
a tibble, or vector of file paths if file is not CSV or if there are multiple
}
\description{
Downloads and reads dataset identified by \code{dataset_id}.
Unzips if necessary, but only loads CSV files, otherwise returns the path to the downloaded file.
Converts types of columns where known, e.g. value columns to numeric.
}
\details{
The schema of the dataset is not yet used, so some columns may be mistyped and are by default returned as character vectors.
}
\note{
Do not use this for harvesting datasets from CZSO en masse.
}
\examples{
\donttest{
czso_get_table("110080")
}
}
\seealso{
Other Core workflow: 
\code{\link{czso_get_catalogue}()},
\code{\link{get_table}()}
}
\concept{Core workflow}
