% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{czso_get_table_schema}
\alias{czso_get_table_schema}
\title{Get CZSO table schema}
\usage{
czso_get_table_schema(dataset_id, resource_num = 1)
}
\arguments{
\item{dataset_id}{Dataset ID}

\item{resource_num}{Resource number, typically 1 in CZSO (the default)}
}
\value{
a tibble with a description of the columns in the table.
}
\description{
Retrieves and parses the schema for the table identified by dataset_id and resource_num.
}
\examples{
get_czso_table_schema("110080")
}
\seealso{
Other Additional tools: 
\code{\link{czso_get_dataset_doc}()},
\code{\link{czso_get_dataset_metadata}()},
\code{\link{get_czso_dataset_doc}()},
\code{\link{get_czso_dataset_metadata}()},
\code{\link{get_czso_table_schema}()}
}
\concept{Additional tools}
