% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_catalogue}
\alias{get_catalogue}
\title{Get catalogue of open CZSO datasets}
\usage{
get_catalogue(
  provider = "\\\\u010cesk\\\\u00fd statistick\\\\u00fd \\\\u00fa\\\\u0159ad",
  title_regex = NULL,
  description_regex = NULL
)
}
\arguments{
\item{provider}{DESCRIPTION.}

\item{title_regex}{DESCRIPTION.}

\item{description_regex}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
