% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_catalogue}
\alias{get_catalogue}
\title{Deprecated: Retrieve and read dataset from CZSO}
\usage{
get_catalogue()
}
\value{
a tibble
}
\description{
Deprecated, use \code{get_czso_catalogue()} instead.
}
\examples{
# see `get_czso_catalogue()`
}
