% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_catalogue}
\alias{get_catalogue}
\title{Deprecated: use \code{czso_get_catalogue()} instead}
\usage{
get_catalogue()
}
\value{
a tibble
}
\description{
\lifecycle{soft-deprecated}
}
\examples{
# see `czso_get_catalogue()`
}
