% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_czso_dataset_metadata}
\alias{get_czso_dataset_metadata}
\title{Deprecated, use \code{czso_get_catalogue()} instead.}
\usage{
get_czso_dataset_metadata(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset ID}
}
\value{
a list
}
\description{
\lifecycle{deprecated}
}
\seealso{
Other Additional tools: 
\code{\link{czso_get_dataset_doc}()},
\code{\link{czso_get_dataset_metadata}()},
\code{\link{czso_get_table_schema}()},
\code{\link{get_czso_dataset_doc}()},
\code{\link{get_czso_table_schema}()}
}
\concept{Additional tools}
