% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_table}
\alias{get_table}
\title{Deprecated: use \code{czso_get_table()} instead.}
\usage{
get_table(dataset_id, resource_num = 1, force_redownload = F)
}
\arguments{
\item{dataset_id}{a character. Found in the czso_id column of data frame returned by \code{get_catalogue()}.}

\item{resource_num}{integer. Order of resource in resource list for the given dataset. Defaults to 1, the normal value for CZSO datasets.}

\item{force_redownload}{integer. Whether to redownload data source file even if already cached. Defaults to FALSE.}
}
\value{
a tibble
}
\description{
\lifecycle{soft-deprecated}.
}
\examples{
# see `czso_get_table()`
}
\seealso{
Other Core workflow: 
\code{\link{czso_get_catalogue}()},
\code{\link{czso_get_table}()}
}
\concept{Core workflow}
